+++ sdk/communication/communication-administration/recordings/node/phonenumberadministrationclient_lists_playbacklive/recording_can_list_phone_plans.js
@@ -1,0 +1,47 @@
+let nock = require("nock");
+
+module.exports.hash = "3cdc938169a6df415554f9d5e6aef39b";
+
+module.exports.testInfo = { uniqueName: {}, newDate: {} };
+
+nock("https://endpoint", { encodedQueryParams: true })
+  .get("/administration/phonenumbers/countries/US/phoneplangroups/sanitized/phoneplans")
+  .query(true)
+  .reply(
+    200,
+    {
+      phonePlans: [
+        {
+          phonePlanId: "sanitized",
+          localizedName: "Outbound Only PSTN For User - Geographic",
+          locationType: "Selection",
+          areaCodes: [],
+          capabilities: ["Azure", "OutboundCalling", "UserAssignment", "Geographic"],
+          maximumSearchSize: 20
+        },
+        {
+          phonePlanId: "sanitized",
+          localizedName: "Inbound Only PSTN For User - Geographic",
+          locationType: "Selection",
+          areaCodes: [],
+          capabilities: ["Azure", "InboundCalling", "UserAssignment", "Geographic"],
+          maximumSearchSize: 20
+        }
+      ],
+      nextLink: null
+    },
+    [
+      "Transfer-Encoding",
+      "chunked",
+      "Content-Type",
+      "application/json; charset=utf-8",
+      "MS-CV",
+      "dBoCEDukuEOYgrlJiZt+hQ.0",
+      "X-Processing-Time",
+      "591ms",
+      "X-Azure-Ref",
+      "0XwJ7XwAAAACmAAAStS0+RL34KH7QHdEyWVZSMzBFREdFMDQwOQA5ZmM3YjUxOS1hOGNjLTRmODktOTM1ZS1jOTE0OGFlMDllODE=",
+      "Date",
+      "Mon, 05 Oct 2020 11:24:15 GMT"
+    ]
+  );