+++ sdk/identity/identity/src/credentials/authorizationCodeCredential.ts
@@ -8,6 +8,7 @@ import { TokenCredential, GetTokenOptions, AccessToken } from "@azure/core-http"
 import { IdentityClient, TokenResponse, TokenCredentialOptions } from "../client/identityClient";
 import { CanonicalCode } from "@opentelemetry/api";
 import { credentialLogger, formatSuccess, formatError } from "../util/logging";
+import { getIdentityTokenEndpointSuffix } from "../util/identityTokenEndpoint";
 
 const logger = credentialLogger("AuthorizationCodeCredential");
 
@@ -156,8 +157,9 @@ export class AuthorizationCodeCredential implements TokenCredential {
       }
 
       if (tokenResponse === null) {
+        const urlSuffix = getIdentityTokenEndpointSuffix(this.tenantId);
         const webResource = this.identityClient.createWebResource({
-          url: `${this.identityClient.authorityHost}/${this.tenantId}/oauth2/v2.0/token`,
+          url: `${this.identityClient.authorityHost}/${this.tenantId}/${urlSuffix}`,
           method: "POST",
           disableJsonStringifyOnBody: true,
           deserializationMapper: undefined,