+++ sdk/keyvault/keyvault-secrets/test/internal/userAgent.spec.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT license.
 
 import * as assert from "assert";
-import { SDK_VERSION } from "../../src/generated/utils/constants";
+import { SDK_VERSION } from "../../src/constants";
 import { packageVersion } from "../../src/generated/keyVaultClientContext";
 import { isNode } from "@azure/core-http";
 import path from "path";