+++ sdk/metricsadvisor/ai-metrics-advisor/samples/typescript/src/hooks.ts
@@ -35,7 +35,7 @@ async function createWebHook(client: MetricsAdvisorAdministrationClient) {
   console.log("Creating a new web hook...");
   const hook: WebhookHook = {
     hookType: "Webhook",
-    hookName: "js web hook example" + new Date().getTime().toFixed(),
+    name: "js web hook example" + new Date().getTime().toFixed(),
     description: "description",
     hookParameter: {
       endpoint: "https://httpbin.org/post",
@@ -54,7 +54,7 @@ async function createEmailHook(client: MetricsAdvisorAdministrationClient) {
   console.log("Creating a new email hook...");
   const hook: EmailHook = {
     hookType: "Email",
-    hookName: "js email hook example" + new Date().getTime().toFixed(),
+    name: "js email hook example" + new Date().getTime().toFixed(),
     description: "description",
     hookParameter: { toList: ["test@example.com"] }
   };
@@ -66,7 +66,7 @@ async function createEmailHook(client: MetricsAdvisorAdministrationClient) {
 async function getHook(client: MetricsAdvisorAdministrationClient, hookId: string) {
   console.log(`Retrieving an existing hook for id ${hookId}...`);
   const result = await client.getHook(hookId);
-  console.log(result.hookName);
+  console.log(result.name);
   console.log(result.description);
   console.log(result.admins);
 }