+++ sdk/tables/data-tables/karma.conf.js
@@ -134,13 +134,7 @@ module.exports = function(config) {
     // start these browsers
     // available browser launchers: https://npmjs.org/browse/keyword/karma-launcher
     // 'ChromeHeadless', 'Chrome', 'Firefox', 'Edge', 'IE'
-    browsers: ["Chrome_without_security"],
-    customLaunchers: {
-      Chrome_without_security: {
-        base: "ChromeHeadless",
-        flags: ["--disable-web-security"]
-      }
-    },
+    browsers: ["ChromeHeadless"],
 
     // Continuous Integration mode
     // if true, Karma captures browsers, runs the tests and exits